Added 1.21.6 locs, added tests for 1.20+ locs, updated some rules, updated Advancement cap, general cleanup
Added support for the two Advancements introduced in 1.21.6, "Stay Hydrated!" and "Heart Transplanter".
Added tests for all of the Advancements added in 1.20 and beyond.
Increased the Advancement cap to 136 to account for all of the new Advancements.
Changed the rules for Wax On and Wax Off to account for the existence of Trial Chambers, and updated their unit tests to match.
Made some other minor adjustments.