Updated waxing rules
As of 1.21, Copper Doors no longer require Copper Blocks to craft, providing an easy way of getting Wax On and Wax Off without needing two copies of Progressive Resource Crafting and completely removing the need for those rules.
This updates those locations to now expect the player to craft a Copper Door instead of a Copper Block, as it is always better in all cases to do so.

Notably, this also serves as a form of futureproofing; 1.21.9 adds even more types of copper blocks that can be waxed, most of which also don't require Copper Blocks (the item) in order to obtain. I believe it is overall better to change the rule now, rather than arbitrarily trying to cram everything made with copper into Progressive Resource Crafting 2, which would also apply to Lightning Rods in 1.21.9 and therefore require other rules to change.